Small re-factoring
