+++ mozci/buildapi.py
@@ -19,27 +19,11 @@ from allthethings import valid_builder
 from buildjson import query_buildjson_info
 from platforms import associated_build_job, does_builder_need_files
 
+from mozci.utils.misc import _all_urls_reachable
+
 LOG = logging.getLogger()
 
 BUILDAPI = 'https://secure.pub.build.mozilla.org/buildapi/self-serve'
-
-
-def _public_url(url):
-    ''' If we run the script outside the Release Engineering infrastructure
-        we need to use the public interface rather than the internal one.
-    '''
-    replace_urls = [
-        ("http://pvtbuilds.pvt.build",
-         "https://pvtbuilds"),
-        ("http://tooltool.pvt.build.mozilla.org/build",
-         "https://secure.pub.build.mozilla.org/tooltool/pvt/build")
-    ]
-    for from_, to_ in replace_urls:
-        if url.startswith(from_):
-            new_url = url.replace(from_, to_)
-            LOG.debug("Replacing url %s ->\n %s" % (url, new_url))
-            return new_url
-    return url
 
 
 def _make_request(url, payload, auth):
@@ -49,15 +33,6 @@ def _make_request(url, payload, auth):
     LOG.debug(" - status code: %s" % req.status_code)
     LOG.debug(" - text:        %s" % BeautifulSoup(req.text).get_text())
     return req
-
-
-def _all_urls_reachable(urls, auth=None):
-    ''' Determine if the URLs are reachable
-    '''
-    for url in urls:
-        url_tested = _public_url(url)
-        LOG.debug("We are going to test if we can reach %s" % url_tested)
-        requests.head(url_tested, auth=auth)
 
 
 def _matching_jobs(buildername, all_jobs):
@@ -214,11 +189,14 @@ def trigger_job(repo_name, revision, buildername, auth,
             "revision": revision
         })
         payload['files'] = json.dumps(files)
+
         url = r'''%s/%s/builders/%s/%s''' % (
             BUILDAPI,
             repo_name,
             trigger,
-            revision)
+            revision
+        )
+
         if not dry_run:
             return _make_request(url, payload, auth)
         else:
@@ -232,7 +210,7 @@ def trigger_job(repo_name, revision, buildername, auth,
 
 
 #
-# Query type of functions
+# Functions to query
 #
 def jobs_running_url(repo_name, revision):
     ''' Returns url of where a developer can login to see the
